The workflow

Outline

PrepareTable is an mutable structure with its field pointing to the source data, configurations, models and train-test results

its contents can be changed by train!, test! or traintest!

List available datasets and load one

List available datasets

datasets()

1×9 DataFrame

Row	PackageName	Dataset	Title	Rows	Columns	Description	TimeStamp	RawData	ZippedData

String15	String15	String	Int64	Int64	Missing	Date	String	String31

1	NCUWiseLab	ARI_G2F820	Data [ARI_G2F820] of [NCUWiseLab]	5832	19	missing	2023-02-15	data/raw/ARI_G2F820_202201-08_20230206.csv	data/NCUWiseLab/ARI_G2F820.gz

Load one as DataFrame

df = dataset("NCUWiseLab", "ARI_G2F820")

5×19 DataFrame

Row	year	month	day	hour	soil_water_content_10cm	soil_water_content_20cm	soil_water_content_30cm	soil_water_content_50cm	soil_water_content_100cm	pressure_G2F820	air_temperature_G2F820	humidity_G2F820	windspeed_avg_G2F820	windspeed_max_G2F820	solar_hours_G2F820	solar_radiation_MJm-2_G2F820	precipitation_G2F820	precipitation	precipitation_max

Int64	Int64	Int64	Int64	Float64	Float64	Float64	Float64	Float64	Float64	Float64	Float64	Float64	Float64	Float64	Float64	Float64	Float64	Float64

1	2022	1	1	0	NaN	NaN	NaN	NaN	NaN	1016.1	14.1	96.0	1.3	2.4	0.0	0.0	0.0	NaN	0.0

2	2022	1	1	1	NaN	NaN	NaN	NaN	NaN	1016.2	13.7	97.0	0.8	3.0	0.0	0.0	0.0	NaN	0.0

3	2022	1	1	2	NaN	NaN	NaN	NaN	NaN	1015.7	13.8	99.0	2.1	2.5	0.0	0.0	0.0	NaN	0.0

4	2022	1	1	3	NaN	NaN	NaN	NaN	NaN	1015.9	13.8	98.0	0.8	3.4	0.0	0.0	0.0	NaN	0.0

5	2022	1	1	4	NaN	NaN	NaN	NaN	NaN	1015.7	13.9	99.0	0.3	1.2	0.0	0.0	0.0	NaN	0.0

Create the PrepareTable object

Use PrepareTableDefault to create an PrepareTable with default configurations:

SWCForecastBase.PrepareTableDefault — Function

Default data processing:

PrepareTableDefault(df::DataFrame) = PrepareTable(df, ConfigPreprocess(), ConfigAccumulate(), ConfigSeriesToSupervised())

or use PrepareTable for general purposes.

SWCForecastBase.PrepareTable — Type

Constructor

PrepareTable(table) = new(table, PrepareTableConfig[])

Field

table::DataFrame
configs::Vector{<:PrepareTableConfig}
status::Union{TrainTestState, Nothing}
supervised_tables::Union{SeriesToSupervised, Nothing}
cache::Cache

Example

    PrepareTable(df::DataFrame, ConfigPreprocess(), ConfigAccumulate(), ConfigSeriesToSupervised())

or simply

    DefaultPrepareTable(df)

See also DefaultPrepareTable.

Configuration

Currently available configurations:

SWCForecastBase.ConfigAccumulate — Type

ConfigAccumulate generate derived variables as the new columns, by accumulating each variable in variables for every interval in each intervals.

Use keyword arguments to construct the object.

Example

ConfigAccumulate(; variables = Cols(:precipitation_max),
                                intervals = [1, 12, 24, 48, 36],
                                unit = "hr"
        )

variables: Cols column selector for selecting columns (variables) to be derived.

intervals: the window length for accumulating an variable.

unit: unit of intervals as appended string of the new column, e.g., day.

SWCForecastBase.ConfigPreprocess — Type

ConfigPreprocess controls the primary feature selection and how the data being preprocessed before training.

Use keyword arguments to construct the object.

Example

ConfigPreprocess(;
        timeargs = Cols(:year, :month, :day, :hour),
        input_features  = Cols(r"air_temp", r"humidity", r"pressure", r"windspeed", r"precipitation"),
        target_features = Cols(r"soil_water_content"),
        preprocessing   = [take_hour_last, removeunreasonables!, imputeinterp!, disallowmissing!, precipmax!],
        )

SWCForecastBase.ConfigSeriesToSupervised — Type

ConfigSeriesToSupervised controls how the data being shifted; it is for preparing the data for supervised model training.

Use keyword arguments to construct the object.

Example

ConfigSeriesToSupervised(;
        shift_x         = [0, -6, -12, -24, -36, -48, -60, -72],
        shift_y         = [1],
        )

shift_x: the time shift for the data using as the input features

shift_y: the time shift for the data using as the target features

Manipulate

SWCForecastBase.preparetable! — Function

preparetable!(PT::PrepareTable, PTC::ConfigPreprocess) generates datetime column by PTC.timeargs, sort! by :datetime, do PTC.preprocessing in @chain and check if the table is continuous in time.

Note

This method will raise essential error, that PTC::ConfigPreprocess should be the first arg in args of PrepareTable(PT, args...). Otherwise, the succeeding processing such as ConfigAccumulate or ConfigSeriesToSupervised may give incorrect results without error.

preparetable!(PT::PrepareTable, PTC::ConfigAccumulate) generates derived variables as new columns. See ConfigAccumulate.

preparetable!(PT::PrepareTable, PTC::ConfigSeriesToSupervised) creates SeriesToSupervised as PT.supervised_tables for training and testing for supervised models.