SWCDatasets

Index

SWCDatasets.DATASET_ABS_DIR

SWCDatasets.abspath

SWCDatasets.dataset

SWCDatasets.dataset

SWCDatasets.dataset_dir

SWCDatasets.dataset_table

SWCDatasets.datasets

SWCDatasets.target_row

SWCDatasets.unzip_file

SWCDatasets.unzip_file

Public interface

SWCDatasets.DATASET_ABS_DIR — Constant

Absolute directory for package SWCDatasets. Also see abspath.

SWCDatasets.abspath — Method

abspath(args...) = joinpath(DATASET_ABS_DIR[], args...)

SWCDatasets.dataset_dir — Method

Absolute path to the "data" of SWCDatasets. See also DATASET_ABS_DIR.

Example

`dataset_dir("doc", "datasets.csv")`

# Output
"~/.julia/dev/SWCDatasets/data/doc/datasets.csv"

SWCDatasets.dataset_table — Method

The path to the index table for datasets in SWCDatasets. If SWCDatasets is added using pkg> dev SWCDatasets in other project/environment, dataset_table() returns "~/.julia/dev/SWCDatasets/src/../data/doc/datasets.csv".

The reason for dataset_table to be a function rather than a constant is that I can redefine it in the scope of test. See test/compdecomp.jl.

SWCDatasets.datasets — Method

datasets() returns the table of this dataset, and define the global variable SWCDatasets.__datasets as this table.

Set ; update_table = true to force update SWCDatasets.__datasets with the dataset_table(); this keyword argument is intended to make some tests can work since in test dataset_table() is mutating. # todo: find a better way to deal with it.

SWCDatasets.target_row — Method

target_row returns the latest information in datasets(). Given package_name, dataset_name, target_row(package_name, dataset_name), target_row returns the last row that matches row.PackageName == package_name && row.Dataset == dataset_name".

SWCDatasets.dataset — Method

dataset(package_name::AbstractString, dataset_name::AbstractString) returns a DataFrame object unzipped from the last row returned by target_row(package_name, dataset_name). This function mimics the dataset function in RDatasets.jl.

SWCDatasets.dataset — Method

dataset(target_path) decompress target_path and returns it as a DataFrame.

Notice!

If you were intended to load target_path under SWCDatasets rather than the current directory your working with, you should apply abspath that target_path = SWCDatasets.abspath(target_path).

SWCDatasets.unzip_file — Method

The same as dataset, but also save the unzip file.

SWCDatasets.unzip_file — Method

unzip_file(target_path) unzip file at target_path to current directory preserve its original name.

Notice!

If you were intended to load target_path under SWCDatasets rather than the current directory your working with, you should apply abspath that target_path = SWCDatasets.abspath(target_path).