Training and Testing

Besides train the model and test it at once, you may have these jobs done separately.

traintest!(PT; 
        train_before = DateTime(2022, 03, 21), 
        test_after = DateTime(2022, 3, 22)) # train and test

is equivalent to

train!(PT; train_before = DateTime(2022, 03, 21))
test!(PT; test_after = DateTime(2022, 3, 22))

See the followings for detailed information:

SWCForecastBase.train! — Function

Training with PrepareTable.

Example

train!(PT::PrepareTable;
        train_before = :auto, # default
        model = manytrees(),    # default
        numpoints_train = 24*120 # default
    )

train_before: The DateTime or Date before which the data is used for training; if :auto, train_before = now().

model: The model; it can be any MLJ model.

numpoints_train: from train_before how many data points (number of rows) to be included for model training.

SWCForecastBase.test! — Function

Example

test!(PT::PrepareTable; test_after = :auto,  # default
    numpoints_test = 480)   # default

test_after: the DateTime or Date after which model prediction (model testing stage) starts. If :auto, test_after = now().

numpoints_test: the number of data points in the "future" to be tested.

SWCForecastBase.traintest! — Function

traintest!(PT::PrepareTable; kwargs...) do train! then test!, accepts all keyword arguments that train! or test! could have.