Library

Index

SWCForecastBase._series2supervised

SWCForecastBase.addcol_accumulation!

SWCForecastBase.cccount

SWCForecastBase.chknnm

SWCForecastBase.convert_types

SWCForecastBase.diffsstable!

SWCForecastBase.fstree

SWCForecastBase.imputemean!

SWCForecastBase.islnan

SWCForecastBase.islnan

SWCForecastBase.isnnm

SWCForecastBase.isoutofrange

SWCForecastBase.manytrees

SWCForecastBase.movingaverage

SWCForecastBase.narrow_types!

SWCForecastBase.removeunreasonables!

SWCForecastBase.series2supervised

SWCForecastBase.twofstree

Public interface

SWCForecastBase._series2supervised — Method

series2supervised(data, range_shift, range_out)

Tansform a time series dataset into a supervised learning dataset.

The features will always be suffixed by an addtional time shift tag "t-i". Also see `splittimetag()andformattime_tag`.

References:

https://machinelearningmastery.com/convert-time-series-supervised-learning-problem-python/

SWCForecastBase.addcol_accumulation! — Method

Add columns that are derived by accumulating corresponding variables using movingaverage.

Example

all_precipstr = names(df, r"precipitation")

apd = Dict( # time intervals to accumulates precipitation
    "1hour" => 6,
    "12hour" => 6*12,
    "1day" => 6*24,
    "2day" => 6*24*2,
    "3day" => 6*24*3
)

addcol_accumulation!(df, all_precipstr, apd)

See also: movingaverage.

SWCForecastBase.cccount — Method

Of a time series ts, cccount(ts) calculate by default the cumulative counts of elements that approximates zero consecutively.

SWCForecastBase.chknnm — Method

chknnm(df) check if DataFrame df contains missing values or NaN. Use this before input df into machine.

SWCForecastBase.convert_types — Method

convert_types(df, column_names_types)

Converts the element type of each column to a user-specified type.

Arguments

df: Dataframe for which you want to convert the eltype of each column

column_names_types: Column names and target types. The type of column_names_types should be able to be unpacked into column names and target types in a for loop.

References

https://discourse.julialang.org/t/how-to-change-field-names-and-types-of-a-dataframe/43991/11

SWCForecastBase.diffsstable! — Method

Of a variable of name varnm, diffsstable!(X0::DataFrame, varnm, tshift) calculates the difference between the non-shifted (suffixed by "_t0") and time-shifted (e.g., "_t-6"), where the difference is the new column for the series-to-supervised table X0.

Example

    (X0,) = series2supervised(...)
    diffsstable!(X0, "precipitation_1hr", -6)

that creates a new column diff6_precipitation_1hr = X0[:, "precipitation_1hr_t0"] .- X0[:, "precipitation_1hr_t-6"].

SWCForecastBase.fstree — Method

Return a composite tree model with FeatureSelector.

SWCForecastBase.imputemean! — Method

imputemean!(df) substitute literal nan values with the statistical means. If all missing for a column, value 999 is imputed.

Notice

You should be aware that imputemean! might does nothing without error message if the input is a view of dataframe (e.g., df[!, Not(:datetime)]).

See also islnan for literal nan.

SWCForecastBase.islnan — Method

Return true if it is literally not a number. For example, all(islnan.(["#VALUE!", "nan", "NaN", "Nan", nothing])) is true.

SWCForecastBase.islnan — Method

For x being the type other than the types listed above, islnan(x) falls back to isnnm(x).

See isnnm.

SWCForecastBase.isnnm — Method

Check if x is missing, nothing or NaN. Different from isnan, for x being either of Nothing, Missing, DateTime, and AbstractString, islnan(x) returns true for Nothing and Missing, and returns false for the rest.

The difference between islnan and isnnm is that, isnnm check only NaN for Not-a-Number. If you input something like "#VALUE!", "NaN", it returns false (NOT missing, nothing or NaN).

SWCForecastBase.isoutofrange — Method

Return true if value is out of the interval between l0 and l1. Noted that if value is missing, nothing, or literally nan (see islnan), it returns false (NOT out-of-range).

SWCForecastBase.manytrees — Method

A random forest using EnsembleModel

SWCForecastBase.movingaverage — Method

movingaverage(v, n) returns a vector of moving averaged v by every n.

SWCForecastBase.narrow_types! — Method

narrow_types!(df)

Narrows the eltype of each column to the type that actually exists in the each column of dataframe.

Arguments

df: Dataframe for which you want to narrow the eltype of each column

References

https://discourse.julialang.org/t/how-to-change-field-names-and-types-of-a-dataframe/43991/9

SWCForecastBase.removeunreasonables! — Method

removeunreasonables!(df_all) convert all column-name specific unreasonable values to missing. Noted that missing, nothing and literal nan is not "unreasonable values".

Noted that removeunreasonables! will NOT deal with literally Not-a-Number value nor raising an error for any literally Not-a-Number value. See isoutofrange, islnan.

SWCForecastBase.series2supervised — Method

To transform a time series dataset into a supervised learning dataset

Example

A = randn(500,20)
df = DataFrame(A, :auto)
X0,y0 = series_to_supervised(df[:,1:end-1], df[:,end])
X1,y1 = series2supervised(
    df[:,1:end-1] => range(-6, -1; step=1),
    df[:,end] => range(0, 0; step=-1)
    )

NOTICE!

The input DataFrame (df) must have complete rows; that is, the corresponding time tag (it might be df.datetime for example) must be consecutive because df is converted to Matrix and shifted using lag.

This function filter the dataframe using completecases.

References:

https://machinelearningmastery.com/convert-time-series-supervised-learning-problem-python/

TODO: write test for series2supervised, by making sure the datetime shift is correct (e.g., "datetimet0" should always be 1 hour ahead of "datetimet-6" for a 10-minute sampling data).

SWCForecastBase.twofstree — Method

Return a composite tree model with TWO FeatureSelector: selector_1 and selector_2.