SWCForecastBase

Index

SWCForecastBase.Cache

SWCForecastBase.ConfigAccumulate

SWCForecastBase.ConfigPreprocess

SWCForecastBase.ConfigSeriesToSupervised

SWCForecastBase.PrepareTable

SWCForecastBase.PrepareTable

SWCForecastBase.SeriesToSupervised

SWCForecastBase.TrainTestState

SWCForecastBase.PrepareTableDefault

SWCForecastBase.PrepareTableDefault

SWCForecastBase._series2supervised

SWCForecastBase.addcol_accumulation!

SWCForecastBase.cccount

SWCForecastBase.convert_types

SWCForecastBase.diffsstable!

SWCForecastBase.fstree

SWCForecastBase.imputemean!

SWCForecastBase.isoutofrange

SWCForecastBase.manytrees

SWCForecastBase.movingaverage

SWCForecastBase.narrow_types!

SWCForecastBase.precipmax!

SWCForecastBase.preparetable!

SWCForecastBase.preparetable!

SWCForecastBase.preparetable!

SWCForecastBase.preparetable!

SWCForecastBase.removeunreasonables!

SWCForecastBase.save

SWCForecastBase.save

SWCForecastBase.save

SWCForecastBase.series2supervised

SWCForecastBase.take_hour_last

SWCForecastBase.test!

SWCForecastBase.test!

SWCForecastBase.train!

SWCForecastBase.train!

SWCForecastBase.traintest!

SWCForecastBase.traintest!

SWCForecastBase.twofstree

Public interface

SWCForecastBase.preparetable! — Method

preparetable!(PT::PrepareTable, PTC::ConfigAccumulate) generates derived variables as new columns. See ConfigAccumulate.

SWCForecastBase.preparetable! — Method

preparetable!(PT::PrepareTable, PTC::ConfigPreprocess) generates datetime column by PTC.timeargs, sort! by :datetime, do PTC.preprocessing in @chain and check if the table is continuous in time.

Note

This method will raise essential error, that PTC::ConfigPreprocess should be the first arg in args of PrepareTable(PT, args...). Otherwise, the succeeding processing such as ConfigAccumulate or ConfigSeriesToSupervised may give incorrect results without error.

SWCForecastBase.preparetable! — Method

preparetable!(PT::PrepareTable, PTC::ConfigSeriesToSupervised) creates SeriesToSupervised as PT.supervised_tables for training and testing for supervised models.

SWCForecastBase.Cache — Type

Field

prepare::Prepare
train::Train
test::Test

See also TrainTestState.

SWCForecastBase.PrepareTable — Method

Given a table::DataFrame and PTCs::PrepareTableConfig..., PrepareTable runs preparetable!(_, PTC::PrepareTableConfig) for PTC in PTCs in @chain.

Example

    PrepareTable(df::DataFrame, ConfigPreprocess(), ConfigSeriesToSupervised())

SWCForecastBase.PrepareTableDefault — Method

Default data processing:

PrepareTableDefault(df::DataFrame) = PrepareTable(df, ConfigPreprocess(), ConfigAccumulate(), ConfigSeriesToSupervised())

SWCForecastBase.save — Method

save(PT::PrepareTable, dir0) save cache in PT to the directory dir0.

Example

dir0 = "MyResults"
save(PT::PrepareTable, dir0)

SWCForecastBase.save — Method

Without specifying the parent folder dir0, a folder of name of randomstring will be created as the parent folder.

Example

save(PT::PrepareTable) # save results to /Tables_Fsd0w4/... for example

SWCForecastBase.test! — Method

Example

test!(PT::PrepareTable; test_after = :auto, numpoints_test = 480)

test_after: the DateTime or Date after which model prediction (model testing stage) starts. If :auto, test_after = now().

numpoints_test: the number of data points in the "future" to be tested.

SWCForecastBase.train! — Method

Training with PrepareTable.

Example

train!(PT::PrepareTable;
        train_before = :auto,
        model = manytrees(),
        numpoints_train = 24*120
    )

train_before: The DateTime or Date before which the data is used for training; if :auto, train_before = now().

model: The model; it can be any MLJ model.

numpoints_train: from train_before how many data points (number of rows) to be included for model training.

SWCForecastBase.traintest! — Method

traintest!(PT::PrepareTable; kwargs...) do train! then test!, accepts all keyword arguments that train! or test! could have.

SWCForecastBase.addcol_accumulation! — Method

Add columns that are derived by accumulating corresponding variables using movingaverage.

Example

all_precipstr = names(df, r"precipitation")

apd = Dict( # time intervals to accumulates precipitation
    "1hour" => 6,
    "12hour" => 6*12,
    "1day" => 6*24,
    "2day" => 6*24*2,
    "3day" => 6*24*3
)

addcol_accumulation!(df, all_precipstr, apd)

See also: movingaverage.

SWCForecastBase.cccount — Method

Of a time series ts, cccount(ts) calculate by default the cumulative counts of elements that approximates zero consecutively.

SWCForecastBase.movingaverage — Method

movingaverage(v, n) returns a vector of moving averaged v by every n.

SWCForecastBase.precipmax! — Method

precipmax!(df::DataFrame) creates precipitation_max by maximize Cols(r"\Aprecipitation") ByRow.

SWCForecastBase.SeriesToSupervised — Type

T is the datetime of Y.

SWCForecastBase._series2supervised — Method

series2supervised(data, range_shift, range_out)

Tansform a time series dataset into a supervised learning dataset.

The features will always be suffixed by an addtional time shift tag "t-i". Also see `splittimetag()andformattime_tag`.

References:

https://machinelearningmastery.com/convert-time-series-supervised-learning-problem-python/

SWCForecastBase.convert_types — Method

convert_types(df, column_names_types)

Converts the element type of each column to a user-specified type.

Arguments

df: Dataframe for which you want to convert the eltype of each column

column_names_types: Column names and target types. The type of column_names_types should be able to be unpacked into column names and target types in a for loop.

References

https://discourse.julialang.org/t/how-to-change-field-names-and-types-of-a-dataframe/43991/11

SWCForecastBase.diffsstable! — Method

Of a variable of name varnm, diffsstable!(X0::DataFrame, varnm, tshift) calculates the difference between the non-shifted (suffixed by "_t0") and time-shifted (e.g., "_t-6"), where the difference is the new column for the series-to-supervised table X0.

Example

    (X0,) = series2supervised(...)
    diffsstable!(X0, "precipitation_1hr", -6)

that creates a new column diff6_precipitation_1hr = X0[:, "precipitation_1hr_t0"] .- X0[:, "precipitation_1hr_t-6"].

SWCForecastBase.narrow_types! — Method

narrow_types!(df)

Narrows the eltype of each column to the type that actually exists in the each column of dataframe.

Arguments

df: Dataframe for which you want to narrow the eltype of each column

References

https://discourse.julialang.org/t/how-to-change-field-names-and-types-of-a-dataframe/43991/9

SWCForecastBase.series2supervised — Method

To transform a time series dataset into a supervised learning dataset

Example

A = randn(500,20)
df = DataFrame(A, :auto)
X0,y0 = series_to_supervised(df[:,1:end-1], df[:,end])
X1,y1 = series2supervised(
    df[:,1:end-1] => range(-6, -1; step=1),
    df[:,end] => range(0, 0; step=-1)
    )

NOTICE!

The input DataFrame (df) must have complete rows; that is, the corresponding time tag (it might be df.datetime for example) must be consecutive because df is converted to Matrix and shifted using lag.

This function filter the dataframe using completecases.

References:

https://machinelearningmastery.com/convert-time-series-supervised-learning-problem-python/

TODO: write test for series2supervised, by making sure the datetime shift is correct (e.g., "datetimet0" should always be 1 hour ahead of "datetimet-6" for a 10-minute sampling data).

SWCForecastBase.imputemean! — Method

imputemean!(df) substitute literal nan values with the statistical means. If all missing for a column, value 999 is imputed.

Notice

You should be aware that imputemean! might does nothing without error message if the input is a view of dataframe (e.g., df[!, Not(:datetime)]).

See also islnan for literal nan.

SWCForecastBase.isoutofrange — Method

Return true if value is out of the interval between l0 and l1. Noted that if value is missing, nothing, or literally nan (see islnan), it returns false (NOT out-of-range).

SWCForecastBase.removeunreasonables! — Method

removeunreasonables!(df_all) convert all column-name specific unreasonable values to missing. Noted that missing, nothing and literal nan is not "unreasonable values".

Noted that removeunreasonables! will NOT deal with literally Not-a-Number value nor raising an error for any literally Not-a-Number value. See isoutofrange, islnan.

SWCForecastBase.take_hour_last — Method

combine the df which were grouped by :hour taking only the last.

SWCForecastBase.fstree — Method

Return a composite tree model with FeatureSelector.

SWCForecastBase.manytrees — Method

A random forest using EnsembleModel

SWCForecastBase.twofstree — Method

Return a composite tree model with TWO FeatureSelector: selector_1 and selector_2.